+++ aplusmaker/app/build.gradle
@@ -4,7 +4,7 @@ android {
     compileSdkVersion 29
     buildToolsVersion "29.0.2"
     defaultConfig {
-        applicationId "com.koreatehc.a_plus_maker"
+        applicationId "com.koreatech.a_plus_maker"
         minSdkVersion 21
         targetSdkVersion 29
         versionCode 1
