+++ aplusmaker/app/src/main/java/com/koreatech/a_plus_maker/GsonHelper.java
@@ -1,4 +1,4 @@
-package com.koreatehc.a_plus_maker;
+package com.koreatech.a_plus_maker;
 
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
