+++ aplusmaker/app/src/main/java/com/koreatech/a_plus_maker/MainActivity.java
@@ -1,11 +1,9 @@
-package com.koreatehc.a_plus_maker;
+package com.koreatech.a_plus_maker;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-
-import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity extends ActivityBase {
