+++ aplusmaker/app/src/main/java/com/koreatech/a_plus_maker/studymode/TTSStudyMode.java
@@ -1,9 +1,9 @@
-package com.koreatehc.a_plus_maker.studymode;
+package com.koreatech.a_plus_maker.studymode;
 
-public class BlinkStudyMode extends StudyModeFactory {
+public class TTSStudyMode extends StudyModeFactory {
     public static final int NORMAL_MAX_LEVEL = 0;
 
-    public BlinkStudyMode(String content) {
+    public TTSStudyMode(String content) {
         super(content, NORMAL_MAX_LEVEL);
     }
