+++ aplusmaker/app/src/test/java/com/koreatehc/a_plus_maker/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.koreatehc.a_plus_maker;
+package com.koreatech.a_plus_maker;
 
 import org.junit.Test;
